+++ package/rhythmicon-rhythm/CHANGES.md
@@ -1,6 +1,11 @@
 # 0.3.0
 
-- Disallow optional `|` before and after pattern
+- Disallow `|...|` pattern strings
+- Add method `includes`
+- Rename method `copy` to `clone`
+- Rename method `isShuffle` to `shuffled`
+- Rename method `rotation` to `rotated`
+- Rewrite README
 
 # 0.2.1